Removed the in-route registering and use ajax instead.
Even if the result is not really accurate and cheating-prone, this way
we bind the even to an user intention to actually read the stuff.

Duplicates are not checking so far. If the user click on more formats
it's an indication that she's really interested, so bonus point.

Also, being a post, the bots are less likely to hit it.