Revert "Ajax preview on edit"
This reverts commit fa494b22823aa7a686eaefbe6a3231e9baaaa86b.

It turned out that there is no real gain (save for the page reload),
as it would need to pull the (possibly edited) body again.

Considering that usually the thing is edited with It's all text, I'm
not sure how it would handle the body changes under its feet.

For now let's revert this and consider it a PoC.