Double check to have the cover file in place
The log files say: ! Unable to load picture or PDF file 'bb-14731490542801882127981.jpg'.

Now, the first guess was a race condition when the user can configure the BB and then
send out the job. However, it turned out that apparently the code is doing the right
thing, i.e. pick the config, copy the coverfile, check if the copying was done, then
proceed, and if the file vanished, continue but without it.

Now, it's possible that the error is due to a bad image. For now added some debug.

Close #81

Please reopen if happens again