Prevent to spawn two instance of the backup app
I suspect that dbc_go execute prerm another time (capturing the
output). Or execute the old one. Or the new one. I don't know. Anyway,
these maintainer's scripts are supposed to be idempotens. So check if
the pidfile already exists and avoid spawning another copy.

As a side comment, this dbconfig-common was a time-drain. Seems to
work, but there is so much which can go south. And honestly looks like
a giant hack. Not happy.

Close #88

I'm sick of this and I hope it works. It does in my testing, but see
above. If not I'll be glad to fix the breakages.