When rebuilding a file, reindex it as well
In theory, we *could* have a bulk job for reindexing only, and we do
have it (used for the git pulling). However, to avoid spamming the UI
with apparently redundant actions, let plug the reindexing into the
rebuilding. And it makes sense, as refreshing the title would also
mean to ensure it's properly indexed. Rebuilding all the site will
thus rebuild not only the formats, but the indexes as well.

The penality is in the order of a couple of seconds per text,
depending how fast the machine is. This is not a big issue because the
bulk job is non-blocking (well, kind of).

This should close #118