Implement the teaser generation
The problems here are (were) multiple: trimming the text as per site
setting in a smart way (without breaking the html), make the links
absolute, so they will not break, do not spam <h[1-6]>, do not break
the HTML with useless id attributes, ignore document internal linking
and add an ellipsis when done.

The implementation *should* be complete. HTML::Element for the win.