Moved static indexes generation to a standalone job
Generating the static indexes is a performance penalty, taking 12-13
seconds for large sites (with 3000 texts). When spawning a total
rebuild job, this multiplies by the number of texts.

Instead, create a rebuild job and when publishing/rebuilding, check if
such job is already scheduled and create it if missing, so we save a
lot of time on large site and on global rebuilding as well.

This appears to work, but probably some unrelated tests (which could
count jobs) are failing.

Close #134