Ensure that on rebuild/reindex we don't lose the standard PDFs
While this is normally not a problem, if we rebuild the whole site the
jobs will have such priority:

 - first all light formats (here the PDFs are nuked by the compiler)
 - then all the PDFs

Depending on the site size, it could be even a whole day while many
titles would be without PDFs.

So save them into the canonical .cXX.pdf which the compiler ignores.