Enforce the standard PDF to have twoside => false
While it makes totally sense to have a custom format which is not
imposed but do have a twoside layout, the standard PDF should be
onesided as it is the behaviour of the compiler when producing imposed
ones, as it did before this release.

Having the imposition embedded in the compiler was definitively a
mistake but can't be helped now (and amusewiki will not use that
feature anymore). It's still useful for the command line, though.