Keep the filters stable
Each filter block is now AND'ed, otherwise makes no sense. Multiple
selections in the same block is allowed, but for this to work, we need
to keep the filter stable.

Say we have three authors. With autosubmit (that's the problem), you
can't select two of them, because once the first is clicked, the form
submits and the filter get applied, cutting out the other two.

So, always retrieve the facets running a search with "no_filters"
turned on, and run a search with "no_facets" turned on, and render the
page combining the output.

With ajax instead of autosubmit, we can save the first run when
refining, as the facets don't change.