Never stack the facets
Otherwise on mobile they become unusable. Well, not that a narrow
column is much better, but it can't be helped. Maybe a modal would be
better, but I'm not really sure.

Also, when the user click, scroll up to the top, where the updated
results are located. Otherwise the user will not get a feeling that
something is actually happening. I think.