Decode the string coming out of Xapian
Search::Xapian doesn't really understand the octects/codepoint stuff.
It's happy to take codepoint as query string and indexing
material (and I believe that's simply because Perl stores them as a
dialect of utf8 internally, so they degrade when passing to a system
library) but when returning them, they are octects, so we need to
decode them.

At some point we shoud encode on indexing and query, but that's why we
have tests in place.