Handle Xapian upgrades asynchronously
The Xapian upgrade could take a long time and blocks the jobber. So,
at jobber start up, ensure that the DBs are up to speed and do the
upgrade. If a new text gets indexed at that time, we lose it, which is
unfortunate.

Please note that a reindex can be done via amusewiki-bootstrap-archive
--xapian so that is cronjob-able to recover the (eventual) missing
texts.