Remove "fw" class or replace it with "fa-fw" as needed
"Fixed width" class for Font Awesome icons is called "fa-fw", not "fw". In most cases it was not necessary or even harmful.

In the navigation bar "fw" is replaced with "fa-fw" to make icons fixed-width. It is important in "mobile" (small width) layout, where icons are stacked vertically, to make the list items align.