script/install_fonts.sh: look through all possible directories for each font
On Debian, as used in current Vagrantfile, /usr/share/texmf/fonts exists, but fonts are actually installed in /usr/share/texlive/texmf-dist/fonts

Therefore, it is necessary to not only to check that directory exists, but that the font directory we are looking for is inside.

Also, /usr/share/texlive/texmf-dist/fonts was not in the list at all.