Install fonts: do not die if fonts are not installed
This script has aged. At the time there wasn't the fontspec.json file,
so these fonts were all you had, and you should have.

So this script is merely an attempt to bring in fonts which are not
visible by fc-list, but that they are installed with texlive
nevertheless.

So development (see: vagrant) is one thing, running is another.

Vagrantfile and debian/control should bring in the fonts needed for
testing, while this script should attempt to install some of them *if*
they are not installed in the font location.

So exiting with an error here is not the correct thing, as the
installation most likely works (unless something else is totally
messed up).