Notify the recipient of the notifications when a text changes status
So far we had:

 - notify when a text is created
 - notify when a change is committed

However, there is no indication of the status change of a
text (usually publication).

I believe this is an useful thing. It could help to catch blunders (a
bit too late maybe, but better than nothing if the mailbox is
monitored).

If this starts spamming, it becomes a problem in itself, but let's
assume for a moment is doing the right thing.

This patch probably breaks mail-related tests (because of the
additional mailings).