Remove the sticky navbar and show it only when scrolling up
Over the years I got really annoyed by the sticky navbar getting in
the way of url #hash fragments, and related hacks which fail randomly.

So try to put an end to this misery: show the sticky navbar only when
scrolling up (back to reach the top) but not when jumping between
anchors.

This should do. There is no smooth transition, though, but it's not so
important I think.