Tweak logic for searching deferred texts
Previously the deferred texts weren't indexed at all, unless the site
option show_preview_when_deferred was true *and* the text had a
teaser.

This is because in some scenario we want to show the teaser
for texts which are not ready for the public, so the search had to
reflect that.

The user of teaser is leveraged on listing pages, which have nothing
to do with a full text search.

On as site which uses show_preview_when_deferred, I think it's fine to
show the texts without teaser in the search results as well. There
will not be much to show, though. I'm reluctant to add another Xapian
value for the teaser boolean just to filter them out.

However, to logged in users, we are now returning always the deferred
texts as well, because as pointed out in #327, user may expect to get
from the search what they see as accessible on the site.

Close #327