Encrypt and check the reset_token
If asked to send reset link, always regenerate the token.

Add method to check it (without creating a password).

The encryption is maybe not-so-needed but given that it comes
basically for free, there is no reasons not to do that, even if the
token lasts just an hour.

App currently in broken/transitional state.