Add comment about the needed TeX live packages
Still wondering about a way to get patches in. I think a stripped down
installation of texlive is doable, but this defeats the purpose of
having an OS integration, which BTW, can't be avoided.

So as long as the fixes are clearly documented, it's probably
acceptable to have some bugs laying around coming from the upstream,
instead of having an unmanageble mess.