Initial implementation for offline editing simplification
This appears to work but it still need:

 1. Way more testing
 2. Update the instructions (do you have a public repo/do you have a
    access to this server?)
 3. Notify committer when a conflict led to archive the tree

The last point shouldn't happen a lot, but it can't be
excluded (hence, at some point it will happen). On a site with low
activity shouldn't happen at all. However, this need a test case
simulating the incident and a mail to the committer(s) should be
fired, telling that their work has been discarded and where to find
it, if needed. This is an unfortunate case, but it's better than
leaving the repos out of sync or in a broken state.

Address #336