Preliminary implementation for automatic toc labels
This is an experimental feature. It is implemented on the front end
side, given that its usage is web-specific (sharing links with a
stable anchor) and appears to work.

It needs to be made opt-in, at least at this stage. And cleaned up a
bit. But I think it should work just fine.