Experiment with the disambiguation
This change is going to be reverted right away. It seems to me that
having a whole procedure for something that should be stable and
predictable is not good, especially for something that should normally
be just a corner case.

Using a parent opens the question: which parent, then depends to its
name for it, as if the parent changes, it changes as well.