RSS feed: do not give the special pages a special treatment
I can't recall why this logic was in place, and looking at it, it does
not make much sense either at this point, as you can control the
pubdate to bump them up.

So let older special entries fall out of the RSS like the rest of the
mortals.

Close #408