Parentage: do not hijack categories
The parent feature, added in 2020, beside displaying the parentage
information on the text page, altered the way the categories were
parsed, so the children would inherit them from the parent.

However, the parentage can be used for cases like antologies,
magazines, etc. where topics and authors should not be overwritten. I
believe this change is not creating much disruption. The only visible
change is that the category listing would show the children as well.

If this is breaking your site, I'm open to restore the feature with an
site option.